fine-tuned-RAG/requirements.txt
langchain>=0.3.0
langchain-openai>=0.3.0
langchain-community>=0.3.0
langchain-chroma>=0.2.0
chromadb>=0.5.0
fastapi>=0.115.0
uvicorn[standard]>=0.34.0
pymupdf>=1.25.0
pdfplumber>=0.11.0
python-dotenv>=1.0.0
tiktoken>=0.8.0
openai>=1.60.0
